Knitr chunk cache fix
When setting cache to true for the chunk labels with knitr the second time knitting the file would break. Since the cache files should've been created but the figures were not properly saved since the directory was manually changed. Was fixed by simply defining the directory for the figures as well.